Fixed Workshops toggling on when purchasing a new one
As far as I understand buildstatus should only trigger once with toggle
and you had been doing it on research (looked at blockyards and
carptentries) but with workshops did it on build which caused them to
toggle back on. At least that is what I figured out hope its right.